+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(splash-screen): update splash screen dimensions and styles; enhance About dialog with logo link. (`73fbd60`)
 - feat(script-refs): introduce normalizeScriptRefsForCompare function and update related tests. (`8c6707c`)
 - chore(dependencies): update @harborclient/sdk to version 1.0.36. (`7efb32e`)
 - feat(about-modal): add social media links to About dialog. (`ef411dd`)